+++ autogpt_platform/frontend/src/components/contextual/PublishAgentModal/components/PublishAgentSelectInfo.tsx
@@ -2,10 +2,10 @@
 
 import * as React from "react";
 import Image from "next/image";
-import { Button } from "../agptui/Button";
-import { IconCross, IconPlus } from "../ui/icons";
+import { Button } from "../../../agptui/Button";
+import { IconCross, IconPlus } from "../../../ui/icons";
 import BackendAPI from "@/lib/autogpt-server-api";
-import { toast } from "../molecules/Toast/use-toast";
+import { toast } from "../../../molecules/Toast/use-toast";
 
 export interface PublishAgentInfoInitialData {
   agent_id: string;
